+++ src/pages/register/register.jsx
@@ -20,12 +20,12 @@ const RegisterPage = () => {
   const history = useHistory();
   const location = useLocation();
   const { isAuth } = useAppSelector((state) => state.auth);
-  const { loginSuccess, refreshUser } = useActions();
+  const { loginSuccess, refreshUser, logout } = useActions();
   const token = document.cookie ? getCookie("token") : "";
   const refreshToken = document.cookie ? getCookie("refreshToken") : "";
   const [
     getUser,
-    { isSuccess: isGetUserSuccess, isError: isGetUserError, data: userData },
+    { isSuccess: isGetUserSuccess, isLoading: isGetUserLoading, isError: isGetUserError, data: userData },
   ] = useLazyGetUserQuery();
   const [
     refreshTokenPost,
@@ -33,16 +33,17 @@ const RegisterPage = () => {
   ] = useRefreshTokenMutation();
 
   useEffect(() => {
-    if (token !== undefined && !isAuth) {
-      console.log(`register`);
+    if (!isGetUserError) {
+      console.log('login')
       getUser(token);
       if (isGetUserSuccess) {
         loginSuccess();
         refreshUser(userData);
       }
-      if (isGetUserError) {
+      if (isGetUserError && refreshToken !== undefined && refreshToken !== '') {
         refreshTokenPost(refreshToken);
         if (isRefreshSuccess) {
+          console.log(isRefreshSuccess);
           let accessToken;
           deleteCookie("refreshToken", token);
           if (userData.accessToken.indexOf("Bearer") === 0) {
@@ -59,7 +60,16 @@ const RegisterPage = () => {
         }
       }
     }
-  }, [token, history, loginSuccess, refreshTokenPost, refreshUser]);
+    if (isGetUserError) {
+      logout();
+      history.replace({
+        pathname: "/login",
+        state: {
+          from: location,
+        },
+      });
+    }
+  }, [getUser, isAuth, isGetUserLoading, isGetUserSuccess, isGetUserError, isRefreshError, isRefreshSuccess, loginSuccess, refreshToken, refreshTokenPost, refreshUser, token, userData]);
 
   const [registerPost] = useRegisterMutation();
